Move config classes to a separate module
